Use test logger
Use test logger implementation in tests for the Program class.
