+++ tests/AdventOfCode.Tests/Puzzles/PuzzleTest.cs
@@ -4,7 +4,6 @@
 namespace MartinCostello.AdventOfCode.Puzzles
 {
     using System;
-    using System.Text;
     using Xunit;
     using Xunit.Abstractions;
 
@@ -63,51 +62,6 @@ namespace MartinCostello.AdventOfCode.Puzzles
             Assert.Equal(0, result);
 
             return puzzle;
-        }
-
-        /// <summary>
-        /// A class representing an implementation of <see cref="ILogger"/> that
-        /// logs to the XUnit output. This class cannot be inherited.
-        /// </summary>
-        private sealed class TestLogger : ILogger
-        {
-            /// <summary>
-            /// Initializes a new instance of the <see cref="TestLogger"/> class.
-            /// </summary>
-            /// <param name="outputHelper">The <see cref="ITestOutputHelper"/> to use.</param>
-            internal TestLogger(ITestOutputHelper outputHelper)
-            {
-                OutputHelper = outputHelper;
-            }
-
-            /// <summary>
-            /// Gets the <see cref="ITestOutputHelper"/> to use.
-            /// </summary>
-            private ITestOutputHelper OutputHelper { get; }
-
-            /// <inheritdoc />
-            public void WriteGrid(bool[,] array, char falseChar, char trueChar)
-            {
-                var builder = new StringBuilder().AppendLine();
-
-                for (int x = 0; x < array.GetLength(0); x++)
-                {
-                    for (int y = 0; y < array.GetLength(1); y++)
-                    {
-                        builder.Append(array[x, y] ? trueChar : falseChar);
-                    }
-
-                    builder.AppendLine();
-                }
-
-                builder.AppendLine();
-
-                OutputHelper.WriteLine(builder.ToString());
-            }
-
-            /// <inheritdoc />
-            public void WriteLine(string format, params object[] args)
-                => OutputHelper.WriteLine(format, args);
         }
     }
 }
